Move Galley helper functions to a Util module
